+++ server/api/projects.js
@@ -878,6 +878,56 @@ const getUserFlaggedProjects = (req, res) => {
 
 
 /**
+ * Retrieve a list of the user's "pinned" projects for quick access.
+ *
+ * @param {Object} req - The express.js request object.
+ * @param {Object} res - The express.js response object.
+ */
+const getUserPinnedProjects = (req, res) => {
+  return User.findOne(
+    { uuid: req.user.decoded.uuid },
+    { pinnedProjects: 1 },
+  ).lean().then((user) => {
+    if (user) {
+      const pinned = user.pinnedProjects || [];
+      return Project.aggregate([
+        {
+          $match: {
+            projectID: {
+              $in: pinned
+            }
+          }
+        }, {
+          $sort: {
+            title: -1,
+          }
+        }, {
+          $project: projectListingProjection
+        }
+      ]);
+    }
+    throw (new Error('user'));
+  }).then((projects) => {
+    return res.send({
+      err: false,
+      projects,
+    });
+  }).catch((err) => {
+    let errMsg = conductorErrors.err6;
+    if (err.message === 'user') {
+      errMsg = conductorErrors.err9;
+    } else {
+      debugError(err);
+    }
+    return res.send({
+      err: false,
+      errMsg,
+    });
+  });
+};
+
+
+/**
  * Retrieves a list of the requesting User's most recent open projects.
  * @param {Object} req - the express.js request object
  * @param {Object} res - the express.js response object
@@ -1520,6 +1570,102 @@ const clearProjectFlag = (req, res) => {
             errMsg: errMsg
         });
     });
+};
+
+
+/**
+ * Looks up whether or not a user has a particular saved in their "pinned" list for quick access.
+ *
+ * @param {Object} req - The express.js request object.
+ * @param {Object} res - The express.js response object.
+ */
+const getProjectPinStatus = (req, res) => {
+  return User.findOne(
+    { uuid: req.user.decoded.uuid },
+    { pinnedProjects: 1 },
+  ).lean().then((user) => {
+    if (user) {
+      let pinned = false;
+      if (Array.isArray(user.pinnedProjects)) {
+        pinned = user.pinnedProjects.includes(req.query.projectID);
+      }
+      return res.send({
+        err: false,
+        pinned,
+      });
+    }
+    throw (new Error('notfound'));
+  }).catch((err) => {
+    let errMsg = conductorErrors.err6;
+    if (err.message === 'notfound') {
+      errMsg = conductorErrors.err9;
+    }
+    return res.send({
+      err: true,
+      errMsg,
+    });
+  });
+};
+
+
+/**
+ * Adds a project to the user's "pinned" list for quick access.
+ *
+ * @param {Object} req - The express.js request object. 
+ * @param {Object} res - The express.js response object.
+ */
+const pinProject = (req, res) => {
+  return Project.findOne({ projectID: req.body.projectID }).lean().then((project) => {
+    if (project) {
+      return User.updateOne({ uuid: req.user.decoded.uuid }, {
+        $addToSet: {
+          pinnedProjects: project.projectID,
+        }
+      });
+    }
+    throw (new Error('notfound'));
+  }).then(() => {
+    return res.send({
+      err: false,
+      msg: 'Project successfully pinned!',
+    });
+  }).catch((err) => {
+    let errMsg = conductorErrors.err6;
+    if (err.message === 'notfound') {
+      errMsg = conductorErrors.err11;
+    } else {
+      debugError(err);
+    }
+    return res.send({
+      err: true,
+      errMsg,
+    });
+  });
+};
+
+
+/**
+ * Removes a project from the user's "pinned" list.
+ *
+ * @param {Object} req - The express.js request object. 
+ * @param {Object} res - The express.js response object.
+ */
+const unpinProject = (req, res) => {
+  return User.updateOne(
+    { uuid: req.user.decoded.uuid },
+    { $pullAll: { pinnedProjects: [req.body.projectID] }},
+  ).then(() => {
+    return res.send({
+      err: false,
+      msg: 'Successfully unpinned project.',
+    });
+  }).catch((err) => {
+    debugError(err);
+    return res.send({
+      err: true,
+      errMsg: conductorErrors.err6,
+    });
+  });
 };
 
 
@@ -2404,6 +2550,18 @@ const validate = (method) => {
             return [
                 body('projectID', conductorErrors.err1).exists().isString().isLength({ min: 10, max: 10 })
             ]
+        case 'getProjectPinStatus':
+            return [
+                query('projectID', conductorErrors.err1).exists().isString().isLength({ min: 10, max: 10 })
+            ]
+        case 'pinProject':
+            return [
+              body('projectID', conductorErrors.err1).exists().isString().isLength({ min: 10, max: 10 }),
+            ]
+        case 'unpinProject':
+            return [
+              body('projectID', conductorErrors.err1).exists().isString().isLength({ min: 10, max: 10 }),
+            ]
         case 'createA11YReviewSection':
             return [
                 body('projectID', conductorErrors.err1).exists().isString().isLength({ min: 10, max: 10 }),
@@ -2442,6 +2600,7 @@ module.exports = {
     getUserProjects,
     getUserProjectsAdmin,
     getUserFlaggedProjects,
+    getUserPinnedProjects,
     getRecentProjects,
     getAvailableProjects,
     getCompletedProjects,
@@ -2452,6 +2611,9 @@ module.exports = {
     removeMemberFromProject,
     flagProject,
     clearProjectFlag,
+    getProjectPinStatus,
+    pinProject,
+    unpinProject,
     notifyProjectCompleted,
     getOrgTags,
     requestProjectPublishing,