+++ server/models/user.js
@@ -37,7 +37,8 @@ const UserSchema = new Schema({
     lastResetAttempt: Date,     // the datetime of the last password reset attempt
     resetToken: String,         // the cryptographically-generated active reset token
     tokenExpiry: Date,          // the datetime that the @resetToken is no longer valid
-    customAvatar: Boolean       // if the user has set their own avatar
+    customAvatar: Boolean,      // if the user has set their own avatar
+    pinnedProjects: [String],   // UUIDs of 'pinned' projects
 }, {
     timestamps: true
 });